Remove a superfluous lifetime parameter
Sure, the struct needs a lifetime parameter, but the arguments to this
function don’t. Given this hypothetical reinstating of the lifetimes:

```rust
pub fn create_post<'a, 'b, 'c>(conn: &'a PgConnection, title: &'b str,
                               body: &'c str) -> Post {
    let new_post: NewPost<'d> = NewPost {
        title: title,
        body: body,
    };
}
```

The question would arise, “what is `'d`?” Answer: it doesn’t escape the
function, so it’s less than `'a` and less than `'b`, so it’s fine. (If
you were to try returning it, *then* you’d need to clarify the title and
body lifetimes.)

Maybe the lifetimes were put there so that *if* you make such a change,
you don’t need to change how the lifetimes are, but it seems unlikely
you’d actually need to do that. Keep things simple until then.