Update examples links
After diesel-rs/diesel#433, we now have the example code in-tree.

All links point to the version of the files from the commit right after
the merge of that PR. I was thinking of maybe changing it to `master`
but I think it's better like this, since it forces us to think about
when we want to update the code snippets.