Fix footer positioning on guide page
Prior to this commit on the guide page the footer would be halfway up
the page which looked weird. Now it's positioned at the bottom of the
page no matter what screen it's on. To do this we give it an absolute
positioning and make the body relative. A couple of tweaks were also
added so that the site still looks the same despite these changes in how
everything is positioned. Most notably still allowing the gradient
change in the background color to be the same on te homepage.