Show usage of `sql_query` on the homepage
On Reddit it was recently mentioned that `sql_query` felt like a second
class API. That was never the intention, and I think it's worth
promoting here.

Either way, I think that `sql_query` and `#[derive(QueryableByName)]` is
much easier to use than the alternatives, even if that's the only part
of Diesel you actually use. We should promote it as such.