Don't pin versions in quickstart
If you use the code as is, you'll get pinned to `1.0.0` forever, which is likely not what most people want.

Alternatively, you could update this document regularly as new versions get released, but that also seems painful. This way, you only have to upgrade it once a new major version gets released.